+++ src/Context.php
@@ -8,23 +8,20 @@ class Context extends BaseContext
 {
     public array $items = [];
 
-    public function append(mixed $item): void
+    public function append(mixed $item): mixed
     {
         $this->items[] = $item;
+        return $item;
     }
 
     public function single(string $name, string|array|null $params = []): CustomSingleDirective
     {
-        $directive = new CustomSingleDirective($name, $params);
-        $this->items[] = $directive;
-        return $directive;
+        return $this->append(new CustomSingleDirective($name, $params));
     }
 
     public function block(string $name, string|array|null $params = []): CustomBlockDirective
     {
-        $directive = new CustomBlockDirective($name, $params);
-        $this->items[] = $directive;
-        return $directive;
+        return $this->append(new CustomBlockDirective($name, $params));
     }
 
     protected function getItems(): array