+++ CHANGELOG.md
@@ -1,3 +1,5 @@
 ## dev-master
 
+* #1: `append($item)` methods return `$item` for chains
+
 ## 0.1.0 (11 June 2021)